[refactor][s]: moved getMetadata function to common utils - refs #7
