+++ lib/get.js
@@ -23,7 +23,7 @@ module.exports.get = async(pkgid) => {
     process.exit(1)
   }
 
-  const metadata = await getMetadata(publisher, pkg)
+  const metadata = await utils.getMetadata(publisher, pkg)
   const bitStoreUrl = metadata.bitstore_url
   const descriptor = metadata.descriptor
   const filesToDownload = getFilesToDownload(bitStoreUrl, descriptor)
@@ -53,20 +53,6 @@ module.exports.get = async(pkgid) => {
 
 }
 
-
-const getMetadata = async (publisher, package, sUrl=utils.getServerUrl() ) => {
-  let apiUrl = `${sUrl}/api/package/${publisher}/${package}`
-  let res = await axios.get(apiUrl).catch( err => {
-    if (err.response && err.response.status === 404) {
-      console.error('Data Package Not Found')
-      process.exit(1)
-    } else {
-      console.error(err.message)
-      process.exit(1)
-    }
-  })
-  return res.data
-}
 
 const checkDestIsEmpty = (publisher, package) => {
   let dest = path.join(publisher, package)
@@ -128,7 +114,6 @@ const downloadFile = async (bitStoreUrl, dest, publisher, pkg, bar) => {
 }
 
 
-module.exports.getMetadata = getMetadata
 module.exports.checkDestIsEmpty = checkDestIsEmpty
 module.exports.downloadFile = downloadFile
 module.exports.getFilesToDownload = getFilesToDownload
