+++ lib/info.js
@@ -1,6 +1,5 @@
 const axios = require('axios')
 const urljoin = require('url-join')
-const {getMetadata} = require('./get.js')
 const {customMarked} = require('./utils/marked.js')
 const {logger} = require('./utils/log-handler.js')
 const utils = require('../lib/utils/common')
@@ -8,7 +7,7 @@ const utils = require('../lib/utils/common')
 const getInfo = async (dhpkgid) => {
   const { publisher, pkg } = utils.parseIdentifier(dhpkgid)
 
-  const metadata = await getMetadata(publisher, pkg)
+  const metadata = await utils.getMetadata(publisher, pkg)
   let readme = await getReadme(urljoin(metadata.bitstore_url, 'README.md'))
   let firstParagraphReadme
