+++ test/test_utils.js
@@ -2,14 +2,39 @@ const test = require('ava')
 const chalk = require('chalk')
 const { logger } = require('../lib/utils/log-handler.js')
 const sinon = require('sinon')
+const nock = require('nock')
 const utils = require('../lib/utils/common.js')
 
+let metadata = {
+  "bitstore_url": "https://bits-staging.datapackaged.com/metadata/publisher/package/_v/latest",
+  "descriptor": {
+    "name": "package",
+    "owner": "publisher",
+    "resources": [
+      {
+        "format": "csv",
+        "name": "firstResource",
+        "path": "test/firsts-resource.csv"
+      },
+      {
+        "format": "csv",
+        "name": "secondResource",
+        "url": "https://example.com/data/second-resource.csv"
+      }
+    ]
+  }
+}
 
 test.beforeEach(t => {
   t.context.error = console.error
   t.context.log = console.log
   console.error = sinon.spy()
   console.log = sinon.spy()
+
+  let getMeta = nock('https://staging.datapackaged.com')
+        .persist()
+        .get('/api/package/publisher/package')
+        .reply(200, metadata)
 })
 
 test.afterEach(t => {
@@ -17,6 +42,10 @@ test.afterEach(t => {
   console.log.reset()
   console.error = t.context.error
   console.log = t.context.log
+})
+
+test.after(t => {
+  nock.restore()
 })
 
 test.serial('Error log is working fine', t => {
@@ -79,4 +108,10 @@ test('Uses default server URL if config not found', t => {
   let sUrl = utils.getServerUrl('not/config')
   let expUrl = 'https://staging.datapackaged.com'
   t.is(sUrl, expUrl)
+})
+
+test('Gets bitStoreUrl if publisher and package is fine', async t => {
+  let sUrl = utils.getServerUrl('not/config')
+  let res = utils.getMetadata('publisher', 'package', sUrl)
+  t.deepEqual(await res, metadata)
 })
